Generate float32 data.

+++ docs/conf.py
@@ -99,7 +99,7 @@ starttimes = {
 
 for name, starttime in starttimes.items():
     db = xdas.Database(
-        data=np.random.randn(*shape),
+        data=np.random.randn(*shape, dtype="float32"),
         coords={
             "time": xdas.Coordinate(
                 tie_indices=[0, shape[0] - 1],
